more casting fixes

+++ Marlin/Marlin_main.cpp
@@ -231,7 +231,7 @@ static float peel_speed = 0; //Used by mUVe 3D Peel Control
 static float peel_pause = 0; //Used by mUVe 3D Peel Control
 static float laser_power = 0; //Used by mUVe 3D laser control
 static float laser_ppm = 10; //Used by mUVe 3D laser control - laser pulses per millimeter
-static int laser_pulse = 16; //Used by mUVe 3D laser control - laser pulse width in cycles
+static int8_t laser_pulse = 16; //Used by mUVe 3D laser control - laser pulse width in cycles
 static float offset[3] = {0.0, 0.0, 0.0};
 static bool home_all_axis = true;
 static float feedrate = 1500.0, next_feedrate, saved_feedrate;
@@ -2470,7 +2470,7 @@ void process_commands()
       else {
 		  laser_ppm=10;
 		}
-	  if(code_seen('C')) laser_pulse = (int) code_value();
+	  if(code_seen('C')) laser_pulse = (int8_t) code_value();
 	  else {
 		  laser_pulse=16;
 		}
